requirements.txt
asgiref==3.3.0
<<<<<<< HEAD
=======
beautifulsoup4==4.9.3
bs4==0.0.1
>>>>>>> ee21e8080f57a9d6ebf48d4d6dd5f47805ceff82
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
Django==3.1.3
django-cors-headers==3.5.0
django-filter==2.4.0
<<<<<<< HEAD
=======
django-pandas==0.6.4
>>>>>>> ee21e8080f57a9d6ebf48d4d6dd5f47805ceff82
django-templated-mail==1.1.1
djangorestframework==3.12.1
djangorestframework-simplejwt
djoser==2.1.0
drf-yasg==1.20.0
<<<<<<< HEAD
=======
drfauto==1.5.2
et-xmlfile==1.0.1
gunicorn==20.1.0
>>>>>>> ee21e8080f57a9d6ebf48d4d6dd5f47805ceff82
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
loguru==0.5.3
MarkupSafe==1.1.1
<<<<<<< HEAD
oauthlib==3.1.0
packaging==20.4
psycopg2
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
=======
numpy==1.20.2
oauthlib==3.1.0
openpyxl==3.0.7
packaging==20.4
pandas==1.2.4
psycopg2-binary
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
>>>>>>> ee21e8080f57a9d6ebf48d4d6dd5f47805ceff82
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.4
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
<<<<<<< HEAD
=======
soupsieve==2.2.1
>>>>>>> ee21e8080f57a9d6ebf48d4d6dd5f47805ceff82
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.11
win32-setctime==1.0.3
zipp==3.4.0
